# Conflicts:
#	composer.json
#	composer.lock
#	src/Configuration/Configuration.php

+++ src/Factory/BinaryFileResponseWrapper.php
@@ -5,6 +5,7 @@ namespace FluffyDiscord\RoadRunnerBundle\Factory;
 use Spiral\RoadRunner\Http\Exception\StreamStoppedException;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Kernel;
 
 /**
  * Basically a copy of BinaryFileResponse->sendContent()
@@ -17,35 +18,50 @@ class BinaryFileResponseWrapper
         $response->prepare($request);
 
         $reflectionClass = new \ReflectionClass($response);
+
+        $tempFileObject = null;
+        if(Kernel::MAJOR_VERSION >= 7 && Kernel::MINOR_VERSION >= 1) {
+            $tempFileObject = $reflectionClass->getProperty("tempFileObject")->getValue($response);
+        }
+
         $maxlen = $reflectionClass->getProperty("maxlen")->getValue($response);
         $offset = $reflectionClass->getProperty("offset")->getValue($response);
-        $chunkSize = $reflectionClass->getProperty("chunkSize")->getValue($response);
+
+        $chunkSize = 16 * 1024;
+        if(Kernel::MAJOR_VERSION >= 6) {
+            $chunkSize = $reflectionClass->getProperty("chunkSize")->getValue($response);
+        }
+
         $deleteFileAfterSend = $reflectionClass->getProperty("deleteFileAfterSend")->getValue($response);
 
         try {
             if (!$response->isSuccessful()) {
-                return;
+                return yield "";
             }
 
-            $file = fopen($response->getFile()->getPathname(), "r");
-
-            if ($maxlen === 0) {
-                return;
+            if (0 === $maxlen) {
+                return yield "";
             }
 
-            if ($offset !== 0) {
-                fseek($file, $offset);
+            if ($tempFileObject) {
+                $file = $tempFileObject;
+                $file->rewind();
+            } else {
+                $file = new \SplFileObject($response->getFile()->getPathname(), 'r');
+            }
+
+            if (0 !== $offset) {
+                $file->fseek($offset);
             }
 
             $length = $maxlen;
-            while ($length && !feof($file)) {
+            while ($length && !$file->eof()) {
                 $read = $length > $chunkSize || 0 > $length ? $chunkSize : $length;
 
-                if (false === $data = fread($file, $read)) {
+                if (false === $data = $file->fread($read)) {
                     break;
                 }
-
-                while ("" !== $data) {
+                while ('' !== $data) {
                     try {
                         yield $data;
                     } catch (StreamStoppedException) {
@@ -58,10 +74,8 @@ class BinaryFileResponseWrapper
                     $data = substr($data, $read);
                 }
             }
-
-            fclose($file);
         } finally {
-            if ($deleteFileAfterSend && is_file($response->getFile()->getPathname())) {
+            if (null === $tempFileObject && $deleteFileAfterSend && is_file($response->getFile()->getPathname())) {
                 unlink($response->getFile()->getPathname());
             }
         }
